+++ src/any/index.ts
@@ -1,0 +1,2 @@
+export { xxh32Any } from "./xxh32Any.js";
+export { xxh64Any } from "./xxh64Any.js";