+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from "./xxh32/index.js";
-export * from "./xxh64/index.js";
+export { xxh32 } from "./xxh32.js";
+export { xxh64 } from "./xxh64.js";